refactor(models): 모델 이름 간소화 및 파일 구조 정리 (#105)
- gemma_3_4b.py → text_generation_model.py로 리네이밍
- e5_embeddings.py → embedding_model.py로 리팩터링
- 불필요한 주석 제거 및 코드 간결화
- 모델 관련 파일 구조 통합 및 명확성 개선
