+++ app/database/group_document_builder.py
@@ -1,6 +1,6 @@
 from typing import Dict, Any
 from app.database.chroma_client import get_chroma_client
-from app.models.e5_embeddings import embed
+from app.models.embedding_model import embed
 
 GROUP_COLLECTION = "group-info"
