+++ app/integrations/kafka/kafka_consumer_manager.py
@@ -442,7 +442,7 @@ class KafkaConsumerManager:
                             timestamp=event_data.get('timestamp', [])
                         )
 
-                        from app.models.e5_embeddings import embed
+                        from app.models.embedding_model import embed
                         client = get_chroma_client()
                         col = client.get_or_create_collection("group-info", embedding_function=embed)
                         col.delete(ids=[f"group-{group_id}"])
@@ -465,7 +465,7 @@ class KafkaConsumerManager:
                             add_documents_to_vector_db(docs, "user-activity")
                             logger.info(f"[ChromaDB] Added user {user_data.userId} to group {user_data.groupId}")
                         elif event_type == "GROUP_LEFT":
-                            from app.models.e5_embeddings import embed
+                            from app.models.embedding_model import embed
                             client = get_chroma_client()
                             col = client.get_or_create_collection("user-activity", embedding_function=embed)
                             col.delete(ids=[f"user-{user_data.userId}-{user_data.groupId}"])
