+++ app/router/v2/vector_db.py
@@ -74,7 +74,7 @@ def delete_group_from_chroma_route(payload: GroupDeleteRequest):
     try:
         group_id = payload.groupId
 
-        from app.models.e5_embeddings import embed
+        from app.models.embedding_model import embed
         client = get_chroma_client()
         col = client.get_or_create_collection("group-info", embedding_function=embed)
         col.delete(ids=[f"group-{group_id}"])
@@ -97,7 +97,7 @@ def remove_user_from_chroma_route(payload: UserRemoveRequest):
 
         ids = [f"user-{user_id}-{group_id}"]
 
-        from app.models.e5_embeddings import embed
+        from app.models.embedding_model import embed
         client = get_chroma_client()
         col = client.get_or_create_collection("user-activity", embedding_function=embed)
         col.delete(ids=ids)
@@ -130,7 +130,7 @@ def list_collection_items(
         offset: int = 0
 ):
     try:
-        from app.models.e5_embeddings import embed
+        from app.models.embedding_model import embed
         client = get_chroma_client()
         col = client.get_or_create_collection(name=collection, embedding_function=embed)
         results = col.get(include=["documents", "metadatas"])
@@ -166,7 +166,7 @@ def get_single_document(
         collection: Literal["group-info", "user-activity", "group-synthetic"] = Query(..., description="조회할 컬렉션 이름"),
         id: str = Query(..., description="조회할 문서 ID")
 ):
-    from app.models.e5_embeddings import embed
+    from app.models.embedding_model import embed
     client = get_chroma_client()
     col = client.get_or_create_collection(name=collection, embedding_function=embed)
